refactor: extract deck parser into UpdateDeckMessageParser
