+++ src/edopro/room/domain/states/dueling/DuelingState.ts
@@ -13,7 +13,7 @@ import { Logger } from "../../../../../shared/logger/domain/Logger";
 import { ISocket } from "../../../../../shared/socket/domain/ISocket";
 import { decimalToBytesBuffer } from "../../../../../utils";
 import { Client } from "../../../../client/domain/Client";
-import { UpdateDeckMessageSizeCalculator } from "../../../../deck/application/UpdateDeckMessageSizeCalculator";
+import { UpdateDeckMessageParser } from "../../../../deck/application/UpdateDeckMessageSizeCalculator";
 import { JoinGameMessage } from "../../../../messages/client-to-server/JoinGameMessage";
 import { PlayerInfoMessage } from "../../../../messages/client-to-server/PlayerInfoMessage";
 import { Commands } from "../../../../messages/domain/Commands";
@@ -140,26 +140,8 @@ export class DuelingState extends RoomState {
 	private handleUpdateDeck(message: ClientMessage, room: Room, client: Client): void {
 		this.logger.debug("DUELING: UPDATE_DECK");
 
-		const messageSize = new UpdateDeckMessageSizeCalculator(message.data).calculate();
-		const body = message.data.subarray(0, messageSize);
-		const mainAndExtraDeckSize = body.readUInt32LE(0);
-		const sizeDeckSize = body.readUint32LE(4);
-		const mainDeck: number[] = [];
-		for (let i = 8; i < mainAndExtraDeckSize * 4 + 8; i += 4) {
-			const code = body.readUint32LE(i);
-			mainDeck.push(code);
-		}
-
-		const sideDeck: number[] = [];
-		for (
-			let i = mainAndExtraDeckSize * 4 + 8;
-			i < (mainAndExtraDeckSize + sizeDeckSize) * 4 + 8;
-			i += 4
-		) {
-			const code = body.readUint32LE(i);
-			sideDeck.push(code);
-		}
-
+		const parser = new UpdateDeckMessageParser(message.data);
+		const [mainDeck, sideDeck] = parser.getDeck();
 		const completeIncomingDeck = [...mainDeck, ...sideDeck];
 		const completeCurrentDeck = [
 			...client.deck.main,
