+++ src/edopro/room/domain/states/waiting/WaitingState.ts
@@ -11,7 +11,7 @@ import { TypeChangeClientMessage } from "../../../../../shared/messages/server-t
 import { ISocket } from "../../../../../shared/socket/domain/ISocket";
 import { Client } from "../../../../client/domain/Client";
 import { DeckCreator } from "../../../../deck/application/DeckCreator";
-import { UpdateDeckMessageSizeCalculator } from "../../../../deck/application/UpdateDeckMessageSizeCalculator";
+import { UpdateDeckMessageParser } from "../../../../deck/application/UpdateDeckMessageSizeCalculator";
 import { JoinGameMessage } from "../../../../messages/client-to-server/JoinGameMessage";
 import { PlayerInfoMessage } from "../../../../messages/client-to-server/PlayerInfoMessage";
 import { Commands } from "../../../../messages/domain/Commands";
@@ -228,26 +228,8 @@ export class WaitingState extends RoomState {
 	): Promise<void> {
 		this.logger.debug("WAITING: UPDATE_DECK");
 		player.updatingDeck();
-		const messageSize = new UpdateDeckMessageSizeCalculator(message.data).calculate();
-		const body = message.data.subarray(0, messageSize);
-		const mainAndExtraDeckSize = body.readUInt32LE(0);
-		const sizeDeckSize = body.readUint32LE(4);
-		const mainDeck: number[] = [];
-		for (let i = 8; i < mainAndExtraDeckSize * 4 + 8; i += 4) {
-			const code = body.readUint32LE(i);
-			mainDeck.push(code);
-		}
-
-		const sideDeck: number[] = [];
-		for (
-			let i = mainAndExtraDeckSize * 4 + 8;
-			i < (mainAndExtraDeckSize + sizeDeckSize) * 4 + 8;
-			i += 4
-		) {
-			const code = body.readUint32LE(i);
-			sideDeck.push(code);
-		}
-
+		const parser = new UpdateDeckMessageParser(message.data);
+		const [mainDeck, sideDeck] = parser.getDeck();
 		const deck = await this.deckCreator.build({
 			main: mainDeck,
 			side: sideDeck,
